refactor: domain 패키지 분리, controller 추가
