+++ src/main/kotlin/racingcar/domain/ForwardStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import java.security.SecureRandom
