+++ src/main/kotlin/racingcar/domain/MoveInfo.kt
@@ -1,3 +1,3 @@
-package racingcar.data
+package racingcar.domain
 
 data class MoveInfo(val carName: String, val moveCount: Int)
