+++ src/test/kotlin/racingcar/CarTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import racingcar.domain.Car
 
 class CarTest {
